+++ src/clojure/nrepl/socket.clj
@@ -7,23 +7,13 @@
   (:require
    [clojure.java.io :as io]
    [nrepl.misc :refer [log]]
-   [nrepl.socket.dynamic :refer [get-path]]
    [nrepl.tls :as tls])
   (:import
-   (java.io BufferedInputStream BufferedOutputStream File OutputStream)
-   (java.net InetSocketAddress ProtocolFamily ServerSocket Socket SocketAddress
-             StandardProtocolFamily URI)
+   (java.io BufferedInputStream BufferedOutputStream OutputStream)
+   (java.net InetSocketAddress ServerSocket Socket URI)
    (java.nio ByteBuffer)
-   (java.nio.file Path)
-   (java.nio.channels Channels ClosedChannelException NetworkChannel
-                      ServerSocketChannel SocketChannel)
+   (java.nio.channels Channels ClosedChannelException ServerSocketChannel SocketChannel)
    (javax.net.ssl SSLServerSocket)))
-
-(defmacro find-class [full-path]
-  `(try
-     (Class/forName (name ~full-path))
-     (catch ClassNotFoundException ex#
-       nil)))
 
 ;;; InetSockets (TCP)
 
@@ -48,154 +38,77 @@
 
 ;; Unix domain sockets
 
-(def ^Class junixsocket-address-class
-  (find-class 'org.newsclub.net.unix.AFUNIXSocketAddress))
+;;;; Compatibility block.
+;; Conditionally loads a particular implementation namespace based on
+;; availability of classes on classpath.
 
-(def ^Class junixsocket-server-class
-  (find-class 'org.newsclub.net.unix.AFUNIXServerSocket))
-
-(def ^Class junixsocket-class
-  (find-class 'org.newsclub.net.unix.AFUNIXSocket))
-
-(def ^Class jdk-unix-address-class
-  (find-class 'java.net.UnixDomainSocketAddress))
+(defn- throw-no-sockets []
+  (let [msg "Support for filesystem sockets requires JDK 17+ or a junixsocket dependency"]
+    (log msg)
+    (throw (ex-info msg {:nrepl/kind ::no-filesystem-sockets}))))
 
 (def unix-domain-flavor
   (cond
-    jdk-unix-address-class :jdk
-    (and junixsocket-address-class junixsocket-server-class) :junixsocket
+    (resolve 'java.net.UnixDomainSocketAddress) :jdk
+    (resolve 'org.newsclub.net.unix.AFUNIXSocketAddress) :junixsocket
     :else nil))
 
-(def jdk-unix-address-of
-  (when (= :jdk unix-domain-flavor)
-    (let [addr-of (.getDeclaredMethod jdk-unix-address-class "of"
-                                      (into-array Class [String]))]
-      (fn [path] (.invoke addr-of nil (into-array String [path]))))))
-
-(def junix-address-of
-  (when (= :junixsocket unix-domain-flavor)
-    (try (let [addr-of (.getDeclaredMethod junixsocket-address-class "of"
-                                           (into-array Class [File]))]
-           (fn [path] (.invoke addr-of nil (into-array File [(File. ^String path)]))))
-         (catch NoSuchMethodException _))))
-
-(def ^:private error-msg
-  "Support for filesystem sockets requires JDK 17+ or a junixsocket dependency")
-
-(defn unix-socket-address
-  "Returns a filesystem socket address for the given path string."
-  [^String path]
-  (case unix-domain-flavor
-    :jdk (jdk-unix-address-of path)
-    :junixsocket (junix-address-of path)
-    (do
-      (log error-msg)
-      (throw (ex-info error-msg {:nrepl/kind ::no-filesystem-sockets})))))
-
-(def jdk-unix-server-socket
-  ;; Dynamic because one argument open doesn't exist until jvm 15, nor UNIX
-  ;; until jvm 16.
-  (when (= :jdk unix-domain-flavor)
-    (let [protocol (-> (.getDeclaredField StandardProtocolFamily "UNIX")
-                       (.get StandardProtocolFamily))
-          protocol (into-array ProtocolFamily [protocol])
-          open (.getDeclaredMethod ServerSocketChannel "open"
-                                   (into-array Class [ProtocolFamily]))]
-      #(.invoke open nil protocol))))
-
-(def jdk-unix-socket
-  ;; Dynamic because one argument open doesn't exist until jvm 15, nor UNIX
-  ;; until jvm 16.
-  (when (= :jdk unix-domain-flavor)
-    (let [protocol (-> (.getDeclaredField StandardProtocolFamily "UNIX")
-                       (.get StandardProtocolFamily))
-          protocol (into-array ProtocolFamily [protocol])
-          open (.getDeclaredMethod SocketChannel "open"
-                                   (into-array Class [ProtocolFamily]))]
-      #(.invoke open nil protocol))))
-
-(def junix-server-socket
-  (when (= :junixsocket unix-domain-flavor)
-    (let [make (.getDeclaredMethod junixsocket-server-class "newInstance" nil)]
-      #(.invoke make nil nil))))
-
-(def junix-socket
-  (when (= :junixsocket unix-domain-flavor)
-    (let [make (.getDeclaredMethod junixsocket-class "newInstance" nil)]
-      #(.invoke make nil nil))))
-
 (defn unix-server-socket
-  "Returns a filesystem socket bound to the path if the JDK is version
-  16 or newer or if com.kohlschutter.junixsocket/junixsocket-core can
-  be loaded dynamically.  Otherwise throws the ex-info map
-  {:nrepl/kind ::no-filesystem-sockets}."
-  [^String path]
-  (let [^SocketAddress addr (unix-socket-address path)]
-    (case unix-domain-flavor
-      :jdk
-      (let [sock (jdk-unix-server-socket)]
-        (.bind ^ServerSocketChannel sock addr)
-        (let [^Path path (get-path addr)]
-          (-> path .toFile .deleteOnExit))
-        sock)
-
-      :junixsocket
-      (let [sock (junix-server-socket)]
-        (.bind ^ServerSocket sock addr)
-        (let [^String path (get-path addr)]
-          (-> path File. .deleteOnExit))
-        sock)
-
-      (do
-        (log error-msg)
-        (throw (ex-info error-msg {:nrepl/kind ::no-filesystem-sockets}))))))
+  "Returns a server filesystem socket bound to the path based on available
+  implementation. Throws ex-info map with {:nrepl/kind ::no-filesystem-sockets}
+  if none is available."
+  ^ServerSocketChannel [path]
+  (case unix-domain-flavor
+    :jdk
+    ((requiring-resolve 'nrepl.socket.jdk17/unix-server-socket) path)
+    :junixsocket
+    ((requiring-resolve 'nrepl.socket.junixsocket/unix-server-socket) path)
+    ;; else
+    (throw-no-sockets)))
 
 (defn unix-client-socket
-  "Returns a filesystem socket bound to the path if the JDK is version
-  16 or newer or if com.kohlschutter.junixsocket/junixsocket-core can
-  be loaded dynamically.  Otherwise throws the ex-info map
-  {:nrepl/kind ::no-filesystem-sockets}."
-  [^String path]
-  (let [^SocketAddress addr (unix-socket-address path)]
-    (case unix-domain-flavor
-      :jdk
-      (let [sock (jdk-unix-socket)]
-        (.connect ^SocketChannel sock addr)
-        sock)
+  "Returns a client filesystem socket bound to the path based on available
+  implementation. Throws ex-info map with {:nrepl/kind ::no-filesystem-sockets}
+  if none is available."
+  ^SocketChannel [path]
+  (case unix-domain-flavor
+    :jdk
+    ((requiring-resolve 'nrepl.socket.jdk17/unix-client-socket) path)
+    :junixsocket
+    ((requiring-resolve 'nrepl.socket.junixsocket/unix-client-socket) path)
+    ;; else
+    (throw-no-sockets)))
 
-      :junixsocket
-      (let [sock (junix-socket)]
-        (.connect ^Socket sock addr)
-        sock)
+(defn- get-socket-address-path
+  "Given a socket address instance, return its filesystem path as a
+  string. Depends on loaded Unix socket implementation."
+  [addr]
+  (case unix-domain-flavor
+    :jdk
+    ((requiring-resolve 'nrepl.socket.jdk17/get-socket-address-path) addr)
+    :junixsocket
+    ((requiring-resolve 'nrepl.socket.junixsocket/get-socket-address-path) addr)
+    ;; else
+    (throw-no-sockets)))
 
-      (do
-        (log error-msg)
-        (throw (ex-info error-msg {:nrepl/kind ::no-filesystem-sockets}))))))
+;;;; Common parts.
 
 (defn as-nrepl-uri
   "Constructs an nREPL URI from a server map and transport scheme.
   Takes a map with :server-socket and :host keys (as in nrepl.server/Server)."
   ^URI [{:keys [host server-socket]} transport-scheme]
-  (cond
-    (instance? ServerSocketChannel server-socket)
+  (if (instance? ServerSocketChannel server-socket)
     (URI. (str transport-scheme "+unix")
-          (let [^Path path (get-path (.getLocalAddress ^NetworkChannel server-socket))]
-            (-> path .toAbsolutePath str))
+          (get-socket-address-path
+           (.getLocalAddress ^ServerSocketChannel server-socket))
           nil)
-
-    (and junixsocket-server-class (instance? junixsocket-server-class server-socket))
-    (URI. (str transport-scheme "+unix")
-          (get-path (.getLocalSocketAddress ^ServerSocket server-socket))
-          nil)
-
-    :else
     (URI. (str transport-scheme
                (when (instance? SSLServerSocket server-socket)
                  "s"))
           nil
           host
           (.getLocalPort ^ServerSocket server-socket)
-          nil nil nil))) ;; fragment
+          nil nil nil)))
 
 (defprotocol Acceptable
   (accept [s]
@@ -233,7 +146,7 @@
 ;; hang if we use both Channels/newInputStream and Channels/newOutputStream.
 ;; Read and write deadlock on a shared channel input/output stream lock
 ;; (cf. https://bugs.openjdk.java.net/browse/JDK-4509080).  Verified that this
-;; still happens (via thread dump when hung) with jdk 16.
+;; still happens (via thread dump when hung) with jdk 17.
 
 (defprotocol Writable
   ;; Underscores were added to satisfy clj-kondo